requirements.txt
aiofiles==0.5.0
aiosqlite==0.17.0
aniso8601==7.0.0
asgiref==3.3.4
async-exit-stack==1.0.1
async-generator==1.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
databases==0.4.3
Django==3.2
djangorestframework==3.12.4
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==2.10
inquirer==2.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
orjson==3.5.1
pandas==1.2.3
promise==2.3
psycopg2-binary==2.8.6
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
Rx==1.6.1
six==1.15.0
SQLAlchemy==1.3.24
sqlparse==0.4.1
starlette==0.13.6
typing-extensions==3.7.4.3
ujson==3.2.0
urllib3==1.26.4
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
wincertstore==0.2
